+++ lib/flutter_multiple_loaders.dart
@@ -17,6 +17,8 @@ export 'src/animations/flipping_card_loader.dart';
 export 'src/animations/glowing_loader.dart';
 export 'src/animations/typing_loader.dart';
 export 'src/animations/blinking_loader.dart';
+export 'src/animations/ripple_loader.dart';
+export 'src/animations/heartbeat_loader.dart';
 
 // Export new innovative animations
 export 'src/animations/dna_helix_loader.dart';